@@ -3,35 +3,35 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-        "@apollo/client": "^3.4.16",
-        "@craco/craco": "^6.1.2",
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.3.2",
-        "@testing-library/user-event": "^7.1.2",
-        "classnames": "^2.3.2",
-        "craco-less": "^1.17.1",
-        "framer-motion": "^1.11.0",
-        "graphql": "^16.0.1",
-        "idx": "^2.5.6",
-        "immer": "^7.0.1",
-        "interactjs": "^1.9.8",
-        "prop-types": "^15.7.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "@apollo/client": "^3.10.4",
+        "@craco/craco": "^7.1.0",
+        "@testing-library/jest-dom": "^6.4.5",
+        "@testing-library/react": "^16.0.0",
+        "@testing-library/user-event": "^14.5.2",
+        "classnames": "^2.5.1",
+        "craco-less": "^3.0.1",
+        "framer-motion": "^11.2.10",
+        "graphql": "^16.8.1",
+        "idx": "^3.0.3",
+        "immer": "^10.1.1",
+        "interactjs": "^1.10.27",
+        "prop-types": "^15.8.1",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
         "react-loadable": "^5.5.0",
-        "react-ranger": "^2.0.2",
-        "react-redux": "^7.2.0",
-        "react-responsive": "8.2.0",
-        "react-router": "^5.1.2",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "4.0.3",
-        "redux": "^4.0.5",
+        "react-ranger": "^2.1.0",
+        "react-redux": "^9.1.2",
+        "react-responsive": "10.0.0",
+        "react-router": "^6.23.1",
+        "react-router-dom": "^6.23.1",
+        "react-scripts": "5.0.1",
+        "redux": "^5.0.1",
         "redux-logger": "^3.0.6",
-        "redux-thunk": "^2.3.0",
-        "styled-components": "^5.1.1",
+        "redux-thunk": "^3.1.0",
+        "styled-components": "^6.1.11",
         "subscriptions-transport-ws": "^0.11.0",
-        "two.js": "^0.7.13",
-        "use-immer": "^0.4.0"
+        "two.js": "^0.8.13",
+        "use-immer": "^0.9.0"
     },
     "scripts": {
         "start": "craco start",
@@ -58,9 +58,7 @@
         ]
     },
     "devDependencies": {
-        "autoprefixer": "^9",
-        "postcss": "^7",
-        "tailwindcss": "npm:@tailwindcss/postcss7-compat",
-        "@babel/plugin-proposal-private-property-in-object": "^7.14.5"
+        "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
+        "tailwindcss": "^3.4.4"
     }
 }
