@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react'
 import { useQuery, useMutation, useSubscription } from '@apollo/client'
 import Two from 'two.js'
-import { useHistory } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 
 import Loadable from 'react-loadable'
-import ZUI from 'two.js/extras/jsm/zui'
+import { ZUI } from 'two.js/extras/jsm/zui'
 
 import {
     UPDATE_COMPONENT_INFO,
@@ -19,6 +19,7 @@ import {
     GET_COMPONENTS_FOR_BOARD,
 } from 'schema/subscriptions'
 import Loader from 'components/utils/loader'
+import { updateX1Y1Vertices, updateX2Y2Vertices } from 'utils/updateVertices'
 
 function getComponentSchema(obj, boardId) {
     return {
@@ -39,6 +40,27 @@ function getComponentSchema(obj, boardId) {
         stroke: obj.stroke,
     }
 }
+
+/**
+ * @typedef {Object} elementData
+ * @property {string} boardId
+ * @property {string} componentType
+ * @property {string} fill
+ * @property {Object} handleDeleteComponent //function
+ * @property {number} height
+ * @property {number} id // component's id
+ * @property {boolean} isDummy
+ * @property {boolean} isGroupSelector
+ * @property {Object} itemData // the data which we recieve from database
+ * @property {Object} metadata
+ * @property {number} prevY // the value of prev Y coordinate used to diff
+ * @property {number} prevX // the value of prev X coordinate used to diff
+ * @property {string} textColor
+ * @property {Object} twoJSInstance // the main Two.js instance
+ * @property {number} width
+ * @property {number} x
+ * @property {number} y
+ */
 
 function addZUI(
     props,
@@ -132,7 +154,11 @@ function addZUI(
             let isGroupSelector = false
 
             let path = e.path || (e.composedPath && e.composedPath())
-            // checks for path in event if it contains following element with condition
+            // checks for path in event if it contains following element with id attr which matches with similar two.js children group
+            // and assigns that specific two.js child(group) to shape object
+
+            // wiki: two.js has base level group (scene which hosts all the objects in instance), so it's like two.js group
+            // inside two.js base level group
             path.forEach((item, index) => {
                 console.log(
                     'item?.classList?.value',
@@ -142,24 +168,100 @@ function addZUI(
                         !item?.classList?.value.includes('avoid-dragging') &&
                         item.tagName === 'g'
                 )
+
                 if (item?.classList?.value.includes('avoid-dragging')) {
                     avoidDragging = true
                 }
+
+                // this does not select path or any inner component
+                // since you mention in condition that only pick with "g" tag name, i.e. group element
                 if (
                     item?.classList?.value &&
                     item?.classList?.value.includes('dragger-picker') &&
                     !item?.classList?.value.includes('avoid-dragging') &&
-                    item.tagName === 'g'
+                    item.tagName === 'g' &&
+                    shape == null
                 ) {
-                    console.log('iterating through path', item.id)
+                    console.log('iterating through path', item.id, shape)
                     console.log(
                         'two scene children',
                         two.scene.children,
                         two.scene.children.find((child) => child.id === item.id)
                     )
-                    shape = two.scene.children.find(
-                        (child) => child.id === item.id
-                    )
+
+                    // if its a arrowLine component and check for class
+                    // because we don't want to select top level group id
+                    if (item?.classList?.value.includes('is-line-circle')) {
+                        console.log('is arrowLine circle ')
+                        console.log(
+                            'get attr of parent id',
+                            document
+                                .getElementById(item.id)
+                                .getAttribute('data-parent-id')
+                        )
+                        let parentId = document
+                            .getElementById(item.id)
+                            .getAttribute('data-parent-id')
+
+                        let lineId = document
+                            .getElementById(item.id)
+                            .getAttribute('data-line-id')
+
+                        let direction = document
+                            .getElementById(item.id)
+                            .getAttribute('data-direction')
+
+                        let getParentTwoData = two.scene.children.find(
+                            (child) => child.id === parentId
+                        )
+
+                        console.log(
+                            'getParentTwoData and its children',
+                            getParentTwoData,
+                            getParentTwoData.children
+                        )
+
+                        let getChildTwoData = getParentTwoData.children.find(
+                            (child) => child.id === item.id
+                        )
+                        let getSiblingChild = getParentTwoData.children.find(
+                            (child) =>
+                                child.id !== item.id &&
+                                child?.children?.length > 0
+                        )
+                        let getLineTwoData = getParentTwoData.children.find(
+                            (child) => child.id === lineId
+                        )
+                        console.log(
+                            'getChildTwoData',
+                            getChildTwoData,
+                            getChildTwoData.type,
+                            ' getSiblingChild',
+                            getSiblingChild
+                        )
+                        shape = getChildTwoData
+
+                        // setting custom properties to existing two properties
+                        shape.lineData = getLineTwoData
+                        shape.direction = direction
+                        shape.siblingCircle = getSiblingChild
+
+                        shape.opacity = 1
+                        shape.siblingCircle.opacity = 1
+
+                        shape.elementData = {
+                            // ...shape.elementData,
+                            isGroupSelector: false,
+                            isLineCircle: true,
+                            lineData: getLineTwoData,
+                            parentData: getParentTwoData,
+                        }
+                        // let getChildTwoData =
+                    } else {
+                        shape = two.scene.children.find(
+                            (child) => child.id === item.id
+                        )
+                    }
                 }
             })
 
@@ -174,6 +276,7 @@ function addZUI(
             // then only perform the mutation
 
             console.log('props.selectPanMode', props.selectPanMode)
+            // in case if it's a group selector, it falls under below condition
             if (shape === null) {
                 // shape = two.scene
                 const { x1, x2, y1, y2 } = {
@@ -222,7 +325,12 @@ function addZUI(
     }
 
     function mousemove(e) {
-        console.log('mouse move event', shape?.elementData)
+        console.log(
+            'mouse move event',
+            shape?.elementData,
+            ' shape lineData',
+            shape?.lineData
+        )
         // console.log('shape in mousemove', e, shape, props.selectPanMode)
         const lastAddedElementId = localStorage.getItem('lastAddedElementId')
         if (lastAddedElementId !== null) {
@@ -244,7 +352,8 @@ function addZUI(
             shape?.id &&
             document.getElementById(shape.id).hasAttribute('data-resize')
         ) {
-            console.log('element resize is being performed')
+            // element resize is being performed
+            // console.log('element resize is being performed')
             domElement.removeEventListener('mousemove', mousemove, false)
             domElement.removeEventListener('mouseup', mouseup, false)
         } else if (shape?.id || shape?.elementData) {
@@ -257,8 +366,62 @@ function addZUI(
                 shape.id !== 'two-0' &&
                 !shape.elementData.isGroupSelector
             ) {
-                shape.position.x += dx / zui.scale
-                shape.position.y += dy / zui.scale
+                // code block condition to handle arrow line circle component's dragging
+                // it roughly translates -> does shape === line/arrow line ?
+                if (shape?.lineData) {
+                    let line = shape?.lineData
+
+                    if (shape.direction === 'left') {
+                        if (e.shiftKey == true) {
+                            let x1 = (line.vertices[0].x += dx / zui.scale)
+                            let y1 = (line.vertices[0].y += dy / zui.scale)
+                            updateX1Y1Vertices(Two, line, x1, y1, shape, two)
+
+                            /* update x2,y2 vertices acc. to shift key press event */
+                            let x2 = line.vertices[1].x
+                            let y2 = y1
+                            updateX2Y2Vertices(
+                                Two,
+                                line,
+                                x2,
+                                y2,
+                                shape.siblingCircle,
+                                two
+                            )
+                        } else {
+                            let x1 = (line.vertices[0].x += dx / zui.scale)
+                            let y1 = (line.vertices[0].y += dy / zui.scale)
+                            updateX1Y1Vertices(Two, line, x1, y1, shape, two)
+                        }
+                    } else if (shape.direction === 'right') {
+                        if (e.shiftKey === true) {
+                            let x2 = (line.vertices[1].x += dx / zui.scale)
+                            let y2 = (line.vertices[1].y += dy / zui.scale)
+                            updateX2Y2Vertices(Two, line, x2, y2, shape, two)
+
+                            /* update x1,y1 vertices acc. to shift key press event */
+                            let x1 = line.vertices[0].x
+                            let y1 = y2
+                            updateX1Y1Vertices(
+                                Two,
+                                line,
+                                x1,
+                                y1,
+                                shape.siblingCircle,
+                                two
+                            )
+                        } else {
+                            let x2 = (line.vertices[1].x += dx / zui.scale)
+                            let y2 = (line.vertices[1].y += dy / zui.scale)
+                            updateX2Y2Vertices(Two, line, x2, y2, shape, two)
+                        }
+                    }
+                }
+                // code block condition to handle normal component's dragging
+                else {
+                    shape.position.x += dx / zui.scale
+                    shape.position.y += dy / zui.scale
+                }
             } else if (shape.elementData.isGroupSelector) {
                 // this blocks falls for the case when user has clicked and
                 // is dragging to create a group selector
@@ -323,21 +486,45 @@ function addZUI(
             ) {
                 // console.log('no need to update')
             } else {
-                // updating already created two.js scene groups to new x,y set
-                shape.elementData.x = shape.translation.x
-                shape.elementData.y = shape.translation.y
+                if (shape?.elementData?.isLineCircle === true) {
+                    console.log('Element is a Line Circle')
+                    // updating already created two.js scene groups to new x,y set
+                    // shape.elementData.x = shape.translation.x
+                    // shape.elementData.y = shape.translation.y
+                    shape.opacity = 0
+                    shape.siblingCircle.opacity = 0
 
-                oldShapeData = { ...shape.elementData }
+                    oldShapeData = { ...shape.elementData }
 
-                newShapeData = Object.assign({}, shape.elementData, {
-                    data: {
-                        x: parseInt(shape.translation.x),
-                        y: parseInt(shape.translation.y),
-                    },
-                })
+                    newShapeData = Object.assign({}, shape.elementData, {
+                        data: {
+                            // x: parseInt(shape.lineData.translation.x),
+                            // y: parseInt(shape.lineData.translation.y),
+                            x1: parseInt(shape.lineData.vertices[0].x),
+                            y1: parseInt(shape.lineData.vertices[0].y),
+                            x2: parseInt(shape.lineData.vertices[1].x),
+                            y2: parseInt(shape.lineData.vertices[1].y),
+                        },
+                    })
 
-                if (shape.elementData.componentType !== 'groupobject') {
                     updateToGlobalState(newShapeData, oldShapeData)
+                } else {
+                    // updating already created two.js scene groups to new x,y set
+                    shape.elementData.x = shape.translation.x
+                    shape.elementData.y = shape.translation.y
+
+                    oldShapeData = { ...shape.elementData }
+
+                    newShapeData = Object.assign({}, shape.elementData, {
+                        data: {
+                            x: parseInt(shape.translation.x),
+                            y: parseInt(shape.translation.y),
+                        },
+                    })
+
+                    if (shape.elementData.componentType !== 'groupobject') {
+                        updateToGlobalState(newShapeData, oldShapeData)
+                    }
                 }
             }
         }
@@ -842,6 +1029,23 @@ const Canvas = (props) => {
                     updateObj: {
                         x: parseInt(newShapeData.data.x),
                         y: parseInt(newShapeData.data.y),
+                        updatedBy: userId,
+                    },
+                },
+            })
+        }
+
+        if (newShapeData?.isLineCircle === true) {
+            updateComponentInfo({
+                variables: {
+                    id: newShapeData.parentData.elementData.id,
+                    updateObj: {
+                        // x: parseInt(newShapeData.data.x),
+                        // y: parseInt(newShapeData.data.y),
+                        x1: parseInt(newShapeData.data.x1),
+                        x2: parseInt(newShapeData.data.x2),
+                        y1: parseInt(newShapeData.data.y1),
+                        y2: parseInt(newShapeData.data.y2),
                         updatedBy: userId,
                     },
                 },
