@@ -1,11 +1,7 @@
 const CracoLessPlugin = require('craco-less')
 
 module.exports = {
-    style: {
-        postcss: {
-            plugins: [require('tailwindcss'), require('autoprefixer')],
-        },
-    },
+    style: {},
     plugins: [
         // {
         //     plugin: CracoLessPlugin,
