@@ -9,6 +9,8 @@ import { UPDATE_COMPONENT_INFO } from 'schema/mutations'
 import Toolbar from 'components/floatingToolbar'
 import ElementCreator from 'factory/arrowLine'
 
+import { updateX1Y1Vertices, updateX2Y2Vertices } from 'utils/updateVertices'
+
 function ArrowLine(props) {
     const [updateComponentInfo] = useMutation(UPDATE_COMPONENT_INFO, {
         ignoreResults: true,
@@ -19,13 +21,12 @@ function ArrowLine(props) {
     const two = props.twoJSInstance
     let selectorInstance = null
     let groupObject = null
-    let resizeLineInstance = null
 
     function onBlurHandler(e) {
-        console.log('on blur arrow line', selectorInstance.pointCircle1.opacity)
+        console.log('on blur arrow line', e)
         // elementOnBlurHandler(e, selectorInstance, two)
-        selectorInstance.pointCircle1.opacity = 0
-        selectorInstance.pointCircle2.opacity = 0
+        selectorInstance.pointCircle1Group.opacity = 0
+        selectorInstance.pointCircle2Group.opacity = 0
         two.update()
         document.getElementById(`${groupObject.id}`) &&
             document
@@ -62,14 +63,20 @@ function ArrowLine(props) {
         // console.log('arrowLine props', props)
         // Instantiate factory
         const elementFactory = new ElementCreator(two, prevX, prevY, {
-            x1: props.x1 || 100,
-            x2: props.x2 || 400,
-            y1: props.y1 || 100,
-            y2: props.y2 || 100,
+            x1: props.x1 || 20,
+            x2: props.x2 || 100,
+            y1: props.y1 || 10,
+            y2: props.y2 || 10,
         })
         // Get all instances of every sub child element
-        const { group, pointCircle1, pointCircle2, resizeLine, line } =
-            elementFactory.createElement()
+        const {
+            group,
+            pointCircle1Group,
+            pointCircle2Group,
+            pointCircle1,
+            pointCircle2,
+            line,
+        } = elementFactory.createElement()
         group.elementData = { ...props.itemData, ...props }
 
         if (props.parentGroup) {
@@ -80,11 +87,11 @@ function ArrowLine(props) {
         } else {
             /** This element will render by creating it's own group wrapper */
             groupObject = group
-            resizeLineInstance = resizeLine
+            // resizeLineInstance = resizeLine
             // const { selector } = getEditComponents(two, group, 4)
             selectorInstance = {
-                pointCircle1,
-                pointCircle2,
+                pointCircle1Group,
+                pointCircle2Group,
             }
 
             two.update()
@@ -95,24 +102,57 @@ function ArrowLine(props) {
             // )
             // some styling
             document.getElementById(line.id).style.cursor = 'pointer'
-            document.getElementById(pointCircle1.id).style.cursor = 'pointer'
-            document.getElementById(pointCircle2.id).style.cursor = 'pointer'
+            document.getElementById(pointCircle1Group.id).style.cursor =
+                'pointer'
+            document.getElementById(pointCircle2Group.id).style.cursor =
+                'pointer'
 
-            document
-                .getElementById(pointCircle1.id)
-                .setAttribute('class', 'avoid-dragging')
-            document
-                .getElementById(pointCircle2.id)
-                .setAttribute('class', 'avoid-dragging')
+            // document
+            //     .getElementById(pointCircle2.id)
+            //     .setAttribute('class', 'dragger-picker')
 
             document
                 .getElementById(group.id)
                 .setAttribute('class', 'dragger-picker')
 
-            // setting database's id in html attribute of element
+            // 1. for point circle 1
             document
-                .getElementById(group.id)
-                .setAttribute('data-component-id', props.id)
+                .getElementById(pointCircle1Group.id)
+                .setAttribute('class', 'dragger-picker is-line-circle')
+
+            document
+                .getElementById(pointCircle1Group.id)
+                .setAttribute('data-parent-id', group.id)
+
+            document
+                .getElementById(pointCircle1Group.id)
+                .setAttribute('data-line-id', line.id)
+
+            document
+                .getElementById(pointCircle1Group.id)
+                .setAttribute('data-direction', 'left')
+
+            // 2. for point circle 2
+            document
+                .getElementById(pointCircle2Group.id)
+                .setAttribute('class', 'dragger-picker is-line-circle')
+
+            document
+                .getElementById(pointCircle2Group.id)
+                .setAttribute('data-parent-id', group.id)
+
+            document
+                .getElementById(pointCircle2Group.id)
+                .setAttribute('data-line-id', line.id)
+
+            document
+                .getElementById(pointCircle2Group.id)
+                .setAttribute('data-direction', 'right')
+
+            // setting database's id in html attribute of element
+            // document
+            //     .getElementById(group.id)
+            //     .setAttribute('data-component-id', props.id)
             // document
             //     .getElementById(group.id)
             //     .setAttribute('data-label', 'line_coord')
@@ -134,13 +174,13 @@ function ArrowLine(props) {
                     id: line.id,
                     data: line,
                 }
-                draft.pointCircle1 = {
-                    id: pointCircle1.id,
-                    data: pointCircle1,
+                draft.pointCircle1Group = {
+                    id: pointCircle1Group.id,
+                    data: pointCircle1Group,
                 }
-                draft.pointCircle2 = {
-                    id: pointCircle2.id,
-                    data: pointCircle2,
+                draft.pointCircle2Group = {
+                    id: pointCircle2Group.id,
+                    data: pointCircle2Group,
                 }
                 draft.text = {
                     data: {},
@@ -155,7 +195,6 @@ function ArrowLine(props) {
             )
             getGroupElementFromDOM.addEventListener('focus', onFocusHandler)
             getGroupElementFromDOM.addEventListener('blur', onBlurHandler)
-
             // const { mousemove, mouseup } = handleDrag(two, group, 'line')
 
             // interact(`#${group.id}`).on('click', () => {
@@ -174,23 +213,21 @@ function ArrowLine(props) {
             interact(`#${group.id}`).on('click', (e) => {
                 console.log('on click group', e)
                 // pointCircle1.opacity = 0
-                resizeLine.opacity = 1
-                pointCircle1.opacity = 1
-                pointCircle1.opacity = 1
-                pointCircle2.opacity = 1
-                pointCircle2.opacity = 1
+                // resizeLine.opacity = 1
+                pointCircle1Group.opacity = 1
+                pointCircle2Group.opacity = 1
 
-                pointCircle1.translation.x = line.vertices[0].x - 0
-                pointCircle1.translation.y =
-                    line.vertices[0].y + parseInt(line.linewidth)
+                // pointCircle1Group.translation.x = line.vertices[0].x - 0
+                // pointCircle1Group.translation.y =
+                //     line.vertices[0].y + parseInt(line.linewidth)
 
                 // here due to x2 being a arrow head, we need some offset.
-                pointCircle2.translation.x =
-                    line.vertices[1].x < line.vertices[0].x
-                        ? line.vertices[1].x - 6
-                        : line.vertices[1].x + 6
-                pointCircle2.translation.y =
-                    line.vertices[1].y + parseInt(line.linewidth)
+                // pointCircle2Group.translation.x =
+                //     line.vertices[1].x < line.vertices[0].x
+                //         ? line.vertices[1].x - 6
+                //         : line.vertices[1].x + 6
+                // pointCircle2Group.translation.y =
+                //     line.vertices[1].y + parseInt(line.linewidth)
                 // selector.update(
                 //     line.getBoundingClientRect(true).left - 10,
                 //     line.getBoundingClientRect(true).right + 10,
@@ -243,136 +280,136 @@ function ArrowLine(props) {
             //     },
             // })
 
-            interact(`#${pointCircle1.id}`).draggable({
-                // enable inertial throwing
-                inertia: false,
+            // interact(`#${pointCircle1.id}`).draggable({
+            //     // enable inertial throwing
+            //     inertia: false,
 
-                listeners: {
-                    start(event) {
-                        console.log(
-                            'start dragging point circle',
-                            event.clientX,
-                            event.pageX,
-                            line.translation.x1,
-                            group.translation.x
-                        )
-                        document
-                            .getElementById(`${pointCircle1.id}`)
-                            .setAttribute('data-resize', 'true')
-                    },
-                    move(event) {
-                        if (event.shiftKey == true) {
-                            let x1 = (line.vertices[0].x += event.dx)
-                            let y1 = (line.vertices[0].y += event.dy)
-                            updateX1Y1Vertices(line, x1, y1, pointCircle1)
+            //     listeners: {
+            //         start(event) {
+            //             console.log(
+            //                 'start dragging point circle',
+            //                 event.clientX,
+            //                 event.pageX,
+            //                 line.translation.x1,
+            //                 group.translation.x
+            //             )
+            //             document
+            //                 .getElementById(`${pointCircle1.id}`)
+            //                 .setAttribute('data-resize', 'true')
+            //         },
+            //         move(event) {
+            //             if (event.shiftKey == true) {
+            //                 let x1 = (line.vertices[0].x += event.dx)
+            //                 let y1 = (line.vertices[0].y += event.dy)
+            //                 updateX1Y1Vertices(line, x1, y1, pointCircle1)
 
-                            /* update x2,y2 vertices acc. to shift key press event */
-                            let x2 = line.vertices[1].x
-                            let y2 = y1
-                            updateX2Y2Vertices(line, x2, y2, pointCircle2)
-                        } else {
-                            let x1 = (line.vertices[0].x += event.dx)
-                            let y1 = (line.vertices[0].y += event.dy)
-                            updateX1Y1Vertices(line, x1, y1, pointCircle1)
-                        }
+            //                 /* update x2,y2 vertices acc. to shift key press event */
+            //                 let x2 = line.vertices[1].x
+            //                 let y2 = y1
+            //                 updateX2Y2Vertices(line, x2, y2, pointCircle2)
+            //             } else {
+            //                 let x1 = (line.vertices[0].x += event.dx)
+            //                 let y1 = (line.vertices[0].y += event.dy)
+            //                 updateX1Y1Vertices(line, x1, y1, pointCircle1)
+            //             }
 
-                        // group.center()
-                        two.update()
-                        // console.log('on move group translation ', group)
-                        // pointCircle1.translation.x =
-                        //     event.pageX + group.translation.x
-                        // pointCircle1.translation.y = event.pageY
-                        // two.update()
-                    },
-                    end(event) {
-                        // Had to use mutation as seperate here in this component due to
-                        // no control of pointcircle available in canvas component
-                        updateComponentInfo({
-                            variables: {
-                                id: props.id,
-                                updateObj: {
-                                    x: group.translation.x,
-                                    y: group.translation.y,
-                                    x1: parseInt(line.vertices[0].x),
-                                    y1: parseInt(line.vertices[0].y),
-                                    x2: parseInt(line.vertices[1].x),
-                                    y2: parseInt(line.vertices[1].y),
-                                },
-                            },
-                        })
-                        document
-                            .getElementById(`${pointCircle1.id}`)
-                            .removeAttribute('data-resize')
-                    },
-                },
-            })
+            //             // group.center()
+            //             two.update()
+            //             // console.log('on move group translation ', group)
+            //             // pointCircle1.translation.x =
+            //             //     event.pageX + group.translation.x
+            //             // pointCircle1.translation.y = event.pageY
+            //             // two.update()
+            //         },
+            //         end(event) {
+            //             // Had to use mutation as seperate here in this component due to
+            //             // no control of pointcircle available in canvas component
+            //             updateComponentInfo({
+            //                 variables: {
+            //                     id: props.id,
+            //                     updateObj: {
+            //                         x: group.translation.x,
+            //                         y: group.translation.y,
+            //                         x1: parseInt(line.vertices[0].x),
+            //                         y1: parseInt(line.vertices[0].y),
+            //                         x2: parseInt(line.vertices[1].x),
+            //                         y2: parseInt(line.vertices[1].y),
+            //                     },
+            //                 },
+            //             })
+            //             document
+            //                 .getElementById(`${pointCircle1.id}`)
+            //                 .removeAttribute('data-resize')
+            //         },
+            //     },
+            // })
 
-            interact(`#${pointCircle2.id}`).draggable({
-                // enable inertial throwing
-                inertia: false,
+            // interact(`#${pointCircle2.id}`).draggable({
+            //     // enable inertial throwing
+            //     inertia: false,
 
-                listeners: {
-                    start(event) {
-                        console.log(
-                            'start dragging point circle',
-                            event.clientX,
-                            event.pageX,
-                            line.translation.x1,
-                            group.translation.x
-                        )
-                        document
-                            .getElementById(`${pointCircle2.id}`)
-                            .setAttribute('data-resize', 'true')
-                    },
-                    move(event) {
-                        if (event.shiftKey === true) {
-                            let x2 = (line.vertices[1].x += event.dx)
-                            let y2 = (line.vertices[1].y += event.dy)
-                            updateX2Y2Vertices(line, x2, y2, pointCircle2)
+            //     listeners: {
+            //         start(event) {
+            //             console.log(
+            //                 'start dragging point circle',
+            //                 event.clientX,
+            //                 event.pageX,
+            //                 line.translation.x1,
+            //                 group.translation.x
+            //             )
+            //             document
+            //                 .getElementById(`${pointCircle2.id}`)
+            //                 .setAttribute('data-resize', 'true')
+            //         },
+            //         move(event) {
+            //             if (event.shiftKey === true) {
+            //                 let x2 = (line.vertices[1].x += event.dx)
+            //                 let y2 = (line.vertices[1].y += event.dy)
+            //                 updateX2Y2Vertices(line, x2, y2, pointCircle2)
 
-                            /* update x1,y1 vertices acc. to shift key press event */
-                            let x1 = line.vertices[0].x
-                            let y1 = y2
-                            updateX1Y1Vertices(line, x1, y1, pointCircle1)
-                        } else {
-                            let x2 = (line.vertices[1].x += event.dx)
-                            let y2 = (line.vertices[1].y += event.dy)
-                            updateX2Y2Vertices(line, x2, y2, pointCircle2)
-                        }
+            //                 /* update x1,y1 vertices acc. to shift key press event */
+            //                 let x1 = line.vertices[0].x
+            //                 let y1 = y2
+            //                 updateX1Y1Vertices(line, x1, y1, pointCircle1)
+            //             } else {
+            //                 let x2 = (line.vertices[1].x += event.dx)
+            //                 let y2 = (line.vertices[1].y += event.dy)
+            //                 updateX2Y2Vertices(line, x2, y2, pointCircle2)
+            //             }
 
-                        // group.center()
-                        two.update()
-                        // console.log(
-                        //     'on move group translation ',
-                        //     group.translation.x
-                        // )
-                        // pointCircle1.translation.x =
-                        //     event.pageX + group.translation.x
-                        // pointCircle1.translation.y = event.pageY
-                        // two.update()
-                    },
-                    end(event) {
-                        // Had to use mutation as seperate here in this component due to
-                        // no control of pointcircle available in canvas component
-                        updateComponentInfo({
-                            variables: {
-                                id: props.id,
-                                updateObj: {
-                                    x: group.translation.x,
-                                    y: group.translation.y,
-                                    x1: parseInt(line.vertices[0].x),
-                                    y1: parseInt(line.vertices[0].y),
-                                    x2: parseInt(line.vertices[1].x),
-                                    y2: parseInt(line.vertices[1].y),
-                                },
-                            },
-                        })
-                        document
-                            .getElementById(`${pointCircle2.id}`)
-                            .removeAttribute('data-resize')
-                    },
-                },
-            })
+            //             // group.center()
+            //             two.update()
+            //             // console.log(
+            //             //     'on move group translation ',
+            //             //     group.translation.x
+            //             // )
+            //             // pointCircle1.translation.x =
+            //             //     event.pageX + group.translation.x
+            //             // pointCircle1.translation.y = event.pageY
+            //             // two.update()
+            //         },
+            //         end(event) {
+            //             // Had to use mutation as seperate here in this component due to
+            //             // no control of pointcircle available in canvas component
+            //             updateComponentInfo({
+            //                 variables: {
+            //                     id: props.id,
+            //                     updateObj: {
+            //                         x: group.translation.x,
+            //                         y: group.translation.y,
+            //                         x1: parseInt(line.vertices[0].x),
+            //                         y1: parseInt(line.vertices[0].y),
+            //                         x2: parseInt(line.vertices[1].x),
+            //                         y2: parseInt(line.vertices[1].y),
+            //                     },
+            //                 },
+            //             })
+            //             document
+            //                 .getElementById(`${pointCircle2.id}`)
+            //                 .removeAttribute('data-resize')
+            //         },
+            //     },
+            // })
 
             // interact(`#${group.id}`).draggable({
             //     // enable inertial throwing
@@ -413,158 +450,169 @@ function ArrowLine(props) {
         }
     }, [])
 
-    const updateX2Y2Vertices = (line, x2, y2, pointCircle2) => {
-        pointCircle2.translation.x =
-            line.vertices[1].x < line.vertices[0].x
-                ? line.vertices[1].x
-                : line.vertices[1].x + 6
-        pointCircle2.translation.y =
-            line.vertices[1].y + parseInt(line.linewidth)
+    // const updateX2Y2Vertices = (line, x2, y2, pointCircle2) => {
+    //     pointCircle2.translation.x =
+    //         line.vertices[1].x < line.vertices[0].x
+    //             ? line.vertices[1].x
+    //             : line.vertices[1].x + 6
+    //     pointCircle2.translation.y =
+    //         line.vertices[1].y + parseInt(line.linewidth)
 
-        // copied code from definition of makeArrow
-        let headlen = 10
+    //     // copied code from definition of makeArrow
+    //     let headlen = 10
 
-        let angle = Math.atan2(y2 - line.vertices[0].y, x2 - line.vertices[0].x)
+    //     let angle = Math.atan2(y2 - line.vertices[0].y, x2 - line.vertices[0].x)
 
-        let vertices = [
-            new Two.Anchor(
-                line.vertices[0].x,
-                line.vertices[0].y,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.move
-            ),
-            new Two.Anchor(
-                x2,
-                y2,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
-            new Two.Anchor(
-                x2 - headlen * Math.cos(angle - Math.PI / 4),
-                y2 - headlen * Math.sin(angle - Math.PI / 4),
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
+    //     let vertices = [
+    //         new Two.Anchor(
+    //             line.vertices[0].x,
+    //             line.vertices[0].y,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.move
+    //         ),
+    //         new Two.Anchor(
+    //             x2,
+    //             y2,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
+    //         new Two.Anchor(
+    //             x2 - headlen * Math.cos(angle - Math.PI / 4),
+    //             y2 - headlen * Math.sin(angle - Math.PI / 4),
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
 
-            new Two.Anchor(
-                x2,
-                y2,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.move
-            ),
-            new Two.Anchor(
-                x2 - headlen * Math.cos(angle + Math.PI / 4),
-                y2 - headlen * Math.sin(angle + Math.PI / 4),
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
-        ]
-        line.vertices = vertices
+    //         new Two.Anchor(
+    //             x2,
+    //             y2,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.move
+    //         ),
+    //         new Two.Anchor(
+    //             x2 - headlen * Math.cos(angle + Math.PI / 4),
+    //             y2 - headlen * Math.sin(angle + Math.PI / 4),
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
+    //     ]
+    //     line.vertices = vertices
 
-        two.update()
-    }
+    //     two.update()
+    // }
 
-    const updateX1Y1Vertices = (line, x1, y1, pointCircle1) => {
-        pointCircle1.translation.x = line.vertices[0].x + 0
-        pointCircle1.translation.y =
-            line.vertices[0].y + parseInt(line.linewidth)
+    // const updateX1Y1Vertices = (line, x1, y1, pointCircle1) => {
+    //     pointCircle1.translation.x = line.vertices[0].x + 0
+    //     pointCircle1.translation.y =
+    //         line.vertices[0].y + parseInt(line.linewidth)
 
-        // copied code from definition of makeArrow
-        let headlen = 10
+    //     // copied code from definition of makeArrow
+    //     let headlen = 10
 
-        let angle = Math.atan2(line.vertices[1].y - y1, line.vertices[1].x - x1)
+    //     let angle = Math.atan2(line.vertices[1].y - y1, line.vertices[1].x - x1)
 
-        let vertices = [
-            new Two.Anchor(
-                x1,
-                y1,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.move
-            ),
-            new Two.Anchor(
-                line.vertices[1].x,
-                line.vertices[1].y,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
-            new Two.Anchor(
-                line.vertices[1].x - headlen * Math.cos(angle - Math.PI / 4),
-                line.vertices[1].y - headlen * Math.sin(angle - Math.PI / 4),
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
+    //     let vertices = [
+    //         new Two.Anchor(
+    //             x1,
+    //             y1,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.move
+    //         ),
+    //         new Two.Anchor(
+    //             line.vertices[1].x,
+    //             line.vertices[1].y,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
+    //         new Two.Anchor(
+    //             line.vertices[1].x - headlen * Math.cos(angle - Math.PI / 4),
+    //             line.vertices[1].y - headlen * Math.sin(angle - Math.PI / 4),
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
 
-            new Two.Anchor(
-                line.vertices[1].x,
-                line.vertices[1].y,
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.move
-            ),
-            new Two.Anchor(
-                line.vertices[1].x - headlen * Math.cos(angle + Math.PI / 4),
-                line.vertices[1].y - headlen * Math.sin(angle + Math.PI / 4),
-                undefined,
-                undefined,
-                undefined,
-                undefined,
-                Two.Commands.line
-            ),
-        ]
-        line.vertices = vertices
+    //         new Two.Anchor(
+    //             line.vertices[1].x,
+    //             line.vertices[1].y,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.move
+    //         ),
+    //         new Two.Anchor(
+    //             line.vertices[1].x - headlen * Math.cos(angle + Math.PI / 4),
+    //             line.vertices[1].y - headlen * Math.sin(angle + Math.PI / 4),
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             undefined,
+    //             Two.Commands.line
+    //         ),
+    //     ]
+    //     line.vertices = vertices
 
-        two.update()
-    }
+    //     two.update()
+    // }
 
     useEffect(() => {
         if (internalState?.group?.data) {
             let groupInstance = internalState.group.data
             let lineInstance = internalState.line.data
-            let pointCircle1 = internalState.pointCircle1.data
-            let pointCircle2 = internalState.pointCircle2.data
-
+            let pointCircle1Group = internalState.pointCircle1Group.data
+            let pointCircle2Group = internalState.pointCircle2Group.data
             groupInstance.translation.x = props.x
             groupInstance.translation.y = props.y
             two.update()
-
-            updateX1Y1Vertices(lineInstance, props.x1, props.y1, pointCircle1)
-            updateX2Y2Vertices(lineInstance, props.x2, props.y2, pointCircle2)
+            // updateX1Y1Vertices(lineInstance, props.x1, props.y1, pointCircle1Group)
+            // updateX2Y2Vertices(lineInstance, props.x2, props.y2, pointCircle2Group)
         }
-    }, [
-        props.x,
-        props.y,
-        props.metadata,
-        props.x1,
-        props.x2,
-        props.y1,
-        props.y2,
-    ])
+    }, [props.x, props.y, props.metadata])
+
+    useEffect(() => {
+        if (internalState?.group?.data) {
+            updateX1Y1Vertices(
+                Two,
+                internalState.line.data,
+                props.x1,
+                props.y1,
+                internalState.pointCircle1Group.data,
+                two
+            )
+            updateX2Y2Vertices(
+                Two,
+                internalState.line.data,
+                props.x2,
+                props.y2,
+                internalState.pointCircle2Group.data,
+                two
+            )
+        }
+    }, [props.x1, props.x2, props.y1, props.y2])
 
     function closeToolbar() {
         toggleToolbar(false)
