@@ -1,5 +1,5 @@
 import Two from 'two.js'
-import ZUI from 'two.js/extras/jsm/zui'
+import { ZUI } from 'two.js/extras/jsm/zui'
 
 import { offsetHeight } from 'constants/misc'
 
